requirements.txt
ai21
python-dotenv
fastapi
uvicorn
httpx
pydantic
openrouter
pytest